refactor v1 expr ref and chain helpers
